use fixed-size integers in Configuration
This is to play nice for Rock log files

+++ src/Configuration.hpp
@@ -34,9 +34,9 @@ namespace imu_aceinna_openimu {
         };
 
         std::string periodic_packet_type = "z1";
-        int periodic_packet_rate = 10;
-        int acceleration_low_pass_filter = 25;
-        int angular_velocity_low_pass_filter = 25;
+        int16_t periodic_packet_rate = 10;
+        int16_t acceleration_low_pass_filter = 25;
+        int16_t angular_velocity_low_pass_filter = 25;
 
         Orientation orientation;
